I made design changes to the current landing page of the Zubhub website.
The design shows how it will look to a non user and to a ner user.
the added enhacments includes-
1. Content changes.
The lanfing page content shoulnot be text heavy for the user to ready.
it should be enganing and short.

2. Font size and boldness changes.
The visuails helps to focus on other parts of the page too.

3. Adding hamburger menu.
Instead of showing the cotacts, about and other options on the top
we can add that in the menu and focus on the searchbar iption as thats
more engaging.

4. Adding an icon to indicate the user is online.
Showing the online state can benefit the user with more engagement.

5. Changing postions of present header assets.
Repostioning of the current header elements to make it more intereactive.

6. Changing sign up and Log in option.
Instead of using the term sign in I changed it into log in. I also changed
the button option to give more attention on getting started and create
new projects as that will redirect to making an account if the user have
no account.

7. Adding a button to explore presents projects.
This will lead to more intereaction and keep th user intrested and also
help the user understand more about how the project works.