switch from pull_request to pull_request_target
pull_request doesn't allow access to github secrets
when a PR is created from a forked repo.
What this means is that the previous version of
auto-create-test-vm github action will only work when
the PR is created from a branch in the main repo and not a fork
(basically it won't work for PRs created by any of the interns.
That defeats the whole purpose of the action)

in contrast pull_request_target does allow access to secrets though we have to use some explicit workarounds.

Issue: #1002
Signed-off-by: Ndibe Raymond Olisaemeka <rolisaemeka-ctr@wikimedia.org>